Change loop index type
The loop index type was uint, but uint type is not defined anywhere (and
I don't find the dependency where it is defined). isize is the new type
used in the loop. Similarly to the other loops, and restored from an old
version of the code (see commit acdb86e)